flask-server/requirements.txt
# pip install -r \requirements.txt

flask
flask-sqlalchemy
flask-bcrypt
python-dotenv


